chore: rename test_skip_db.py -> ci_images_db.py
